% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateWisdom.R
\name{CombineWisdom}
\alias{CombineWisdom}
\title{Combine a set of wisdom files for FFTW}
\usage{
CombineWisdom(Threads = c(2L, 4L, 8L, 16L))
}
\arguments{
\item{Threads}{an integer vector specifying the distinct combination of threads with existing wisdom files}

\item{Execute}{Whether or not the command should be executed or printed to stdout}
}
\description{
Combine a set of wisdom files for FFTW
}
