% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateWisdom.R
\name{GenerateWisdom}
\alias{GenerateWisdom}
\title{Generate a set of wisdom files for FFTW}
\usage{
GenerateWisdom(
  Threads = c(2L, 4L, 8L, 16L),
  Execute = FALSE,
  Mode = "FFTW_PATIENT"
)
}
\arguments{
\item{Threads}{an integer vector specifying the distinct combination of threads desired.}

\item{Execute}{Whether or not the commands should be executed or printed to stdout}
}
\description{
Generate a set of wisdom files for FFTW
}
