% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RfCWT-package.R
\docType{package}
\name{RfCWT-package}
\alias{RfCWT}
\alias{RfCWT-package}
\title{RfCWT: The fast Continuous Wavelet Transform (fCWT)}
\description{
"The spectral analysis of signals is currently either dominated by the speed-accuracy trade-off or ignores a signal's often non-stationary character. Here we introduce an open-source algorithm to calculate the fast continuous wavelet transform (fCWT). The parallel environment of fCWT separates scale-independent and scale-dependent operations, while utilizing optimized fast Fourier transforms that exploit downsampled wavelets. fCWT is benchmarked for speed against eight competitive algorithms, tested on noise resistance and validated on synthetic electroencephalography and in vivo extracellular local field potential data. fCWT is shown to have the accuracy of CWT, to have 100 times higher spectral resolution than algorithms equal in speed, to be 122 times and 34 times faster than the reference and fastest state-of-the-art implementations and we demonstrate its real-time performance, as confirmed by the real-time analysis ratio. fCWT provides an improved balance between speed and accuracy, which enables real-time, wide-band, high-quality, time-frequency analysis of non-stationary noisy signals."
}
\author{
\strong{Maintainer}: Matt Summersgill \email{matt.summersgill@gmail.com} (\href{https://orcid.org/0009-0007-2955-5543}{ORCID})

Authors:
\itemize{
  \item Lukas Arts
  \item Egon van den Broek
}

}
\keyword{internal}
